fix scalastyle
