+++ src/main/scala/io/iohk/ethereum/network/FastSyncHost.scala
@@ -4,7 +4,8 @@ import akka.actor.Actor
 import akka.event.LoggingAdapter
 import akka.util.ByteString
 import io.iohk.ethereum.domain.{BlockHeader, Blockchain}
-import io.iohk.ethereum.network.PeerActor.{PeerConfiguration, RLPxConnection}
+import io.iohk.ethereum.network.PeerActor.RLPxConnection
+import io.iohk.ethereum.network.PeerManagerActor.PeerConfiguration
 import io.iohk.ethereum.network.p2p.messages.PV62.{BlockBodies, BlockHeaders, GetBlockBodies, GetBlockHeaders}
 import io.iohk.ethereum.network.p2p.messages.PV63._
 import io.iohk.ethereum.network.rlpx.RLPxConnectionHandler
